refactor(net): hoist host-pattern matcher to mur-common (single source for egress + gateway)
host_matches_pattern / host_allowed moved verbatim from
mur-agent-runtime/src/sandbox/reqwest_guard.rs into mur-common/src/net.rs.
reqwest_guard.rs now re-exports them so egress_proxy.rs callers are
unchanged. A security boundary must not have two copies of this matcher
that can drift.
