+++ mur-common/src/lib.rs
@@ -39,6 +39,7 @@ pub mod model;
 pub mod model_resolve;
 pub mod multimodal;
 pub mod muragent;
+pub mod net;
 pub mod panel;
 pub mod parallel;
 pub mod parameterize;
